requirements.txt
fastapi==0.100.1
uvicorn[standard]
qiskit-ibm-provider
xanadu-cloud-client
amazon-braket-sdk
python-dotenv